requirements.txt
Flask==3.0.3
gunicorn==22.0.0
openai~=1.35.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
httpx==0.27.0
gevent==24.2.1
Flask-SocketIO==5.3.6
playwright==1.44.0
nest-asyncio==1.6.0
gevent-websocket==0.10.1